Refactor MountTable operations so that storage operations can be done internally by MountTable (using a StoragePacker) instead of by Core on behalf of MountTable.
